+++ crates/api/src/site/mod_log.rs
@@ -70,3 +70,225 @@ pub async fn get_mod_log(
     prev_page,
   }))
 }
+
+#[cfg(test)]
+mod tests {
+  use super::*;
+  use lemmy_api_common::utils::remove_or_restore_user_data;
+  use lemmy_db_schema::{
+    source::{
+      comment::{Comment, CommentInsertForm},
+      community::{Community, CommunityInsertForm},
+      instance::Instance,
+      mod_log::moderator::{ModRemoveComment, ModRemovePost},
+      person::{Person, PersonInsertForm},
+      post::{Post, PostInsertForm},
+    },
+    traits::Crud,
+    ModlogActionType,
+  };
+  use lemmy_db_views_modlog_combined::{
+    impls::ModlogCombinedQuery,
+    ModRemoveCommentView,
+    ModRemovePostView,
+    ModlogCombinedView,
+  };
+  use pretty_assertions::assert_eq;
+  use serial_test::serial;
+
+  #[tokio::test]
+  #[serial]
+  async fn test_mod_remove_or_restore_data() -> LemmyResult<()> {
+    let context = LemmyContext::init_test_context().await;
+    let pool = &mut context.pool();
+
+    let inserted_instance = Instance::read_or_create(pool, "my_domain.tld".to_string()).await?;
+
+    let new_mod = PersonInsertForm::test_form(inserted_instance.id, "modder");
+    let inserted_mod = Person::create(pool, &new_mod).await?;
+
+    let new_person = PersonInsertForm::test_form(inserted_instance.id, "chrimbus");
+    let inserted_person = Person::create(pool, &new_person).await?;
+
+    let new_community = CommunityInsertForm::new(
+      inserted_instance.id,
+      "mod_community crepes".to_string(),
+      "nada".to_owned(),
+      "pubkey".to_string(),
+    );
+    let inserted_community = Community::create(pool, &new_community).await?;
+
+    let post_form_1 = PostInsertForm::new(
+      "A test post tubular".into(),
+      inserted_person.id,
+      inserted_community.id,
+    );
+    let inserted_post_1 = Post::create(pool, &post_form_1).await?;
+
+    let post_form_2 = PostInsertForm::new(
+      "A test post radical".into(),
+      inserted_person.id,
+      inserted_community.id,
+    );
+    let inserted_post_2 = Post::create(pool, &post_form_2).await?;
+
+    let comment_form_1 = CommentInsertForm::new(
+      inserted_person.id,
+      inserted_post_1.id,
+      "A test comment tubular".into(),
+    );
+    let _inserted_comment_1 = Comment::create(pool, &comment_form_1, None).await?;
+
+    let comment_form_2 = CommentInsertForm::new(
+      inserted_person.id,
+      inserted_post_2.id,
+      "A test comment radical".into(),
+    );
+    let _inserted_comment_2 = Comment::create(pool, &comment_form_2, None).await?;
+
+    // Remove the user data
+    remove_or_restore_user_data(
+      inserted_mod.id,
+      inserted_person.id,
+      true,
+      &Some("a remove reason".to_string()),
+      &context,
+    )
+    .await?;
+
+    // Verify that their posts and comments are removed.
+    // Posts
+    let post_modlog = ModlogCombinedQuery {
+      type_: Some(ModlogActionType::ModRemovePost),
+      ..Default::default()
+    }
+    .list(pool)
+    .await?;
+    assert_eq!(2, post_modlog.len());
+
+    assert!(matches!(
+      &post_modlog[..],
+      [
+        ModlogCombinedView::ModRemovePost(ModRemovePostView {
+          mod_remove_post: ModRemovePost { removed: true, .. },
+          post: Post { removed: true, .. },
+          ..
+        }),
+        ModlogCombinedView::ModRemovePost(ModRemovePostView {
+          mod_remove_post: ModRemovePost { removed: true, .. },
+          post: Post { removed: true, .. },
+          ..
+        }),
+      ],
+    ));
+
+    // Comments
+    let comment_modlog = ModlogCombinedQuery {
+      type_: Some(ModlogActionType::ModRemoveComment),
+      ..Default::default()
+    }
+    .list(pool)
+    .await?;
+    assert_eq!(2, comment_modlog.len());
+
+    assert!(matches!(
+      &comment_modlog[..],
+      [
+        ModlogCombinedView::ModRemoveComment(ModRemoveCommentView {
+          mod_remove_comment: ModRemoveComment { removed: true, .. },
+          comment: Comment { removed: true, .. },
+          ..
+        }),
+        ModlogCombinedView::ModRemoveComment(ModRemoveCommentView {
+          mod_remove_comment: ModRemoveComment { removed: true, .. },
+          comment: Comment { removed: true, .. },
+          ..
+        }),
+      ],
+    ));
+
+    // Now restore the content, and make sure it got appended
+    remove_or_restore_user_data(
+      inserted_mod.id,
+      inserted_person.id,
+      false,
+      &Some("a restore reason".to_string()),
+      &context,
+    )
+    .await?;
+
+    // Posts
+    let post_modlog = ModlogCombinedQuery {
+      type_: Some(ModlogActionType::ModRemovePost),
+      ..Default::default()
+    }
+    .list(pool)
+    .await?;
+    assert_eq!(4, post_modlog.len());
+
+    assert!(matches!(
+      &post_modlog[..],
+      [
+        ModlogCombinedView::ModRemovePost(ModRemovePostView {
+          mod_remove_post: ModRemovePost { removed: false, .. },
+          post: Post { removed: false, .. },
+          ..
+        }),
+        ModlogCombinedView::ModRemovePost(ModRemovePostView {
+          mod_remove_post: ModRemovePost { removed: false, .. },
+          post: Post { removed: false, .. },
+          ..
+        }),
+        ModlogCombinedView::ModRemovePost(ModRemovePostView {
+          mod_remove_post: ModRemovePost { removed: true, .. },
+          post: Post { removed: false, .. },
+          ..
+        }),
+        ModlogCombinedView::ModRemovePost(ModRemovePostView {
+          mod_remove_post: ModRemovePost { removed: true, .. },
+          post: Post { removed: false, .. },
+          ..
+        }),
+      ],
+    ));
+
+    // Comments
+    let comment_modlog = ModlogCombinedQuery {
+      type_: Some(ModlogActionType::ModRemoveComment),
+      ..Default::default()
+    }
+    .list(pool)
+    .await?;
+    assert_eq!(4, comment_modlog.len());
+
+    assert!(matches!(
+      &comment_modlog[..],
+      [
+        ModlogCombinedView::ModRemoveComment(ModRemoveCommentView {
+          mod_remove_comment: ModRemoveComment { removed: false, .. },
+          comment: Comment { removed: false, .. },
+          ..
+        }),
+        ModlogCombinedView::ModRemoveComment(ModRemoveCommentView {
+          mod_remove_comment: ModRemoveComment { removed: false, .. },
+          comment: Comment { removed: false, .. },
+          ..
+        }),
+        ModlogCombinedView::ModRemoveComment(ModRemoveCommentView {
+          mod_remove_comment: ModRemoveComment { removed: true, .. },
+          comment: Comment { removed: false, .. },
+          ..
+        }),
+        ModlogCombinedView::ModRemoveComment(ModRemoveCommentView {
+          mod_remove_comment: ModRemoveComment { removed: true, .. },
+          comment: Comment { removed: false, .. },
+          ..
+        }),
+      ],
+    ));
+
+    Instance::delete(pool, inserted_instance.id).await?;
+
+    Ok(())
+  }
+}
